Create an entity for items of each domain with their specific relevant metadata.
