+++ src/java/CdbWebPortal/src/java/gov/anl/aps/cdb/portal/model/db/entities/ItemDomainLocation.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) UChicago Argonne, LLC. All rights reserved.
+ * See LICENSE file.
+ */
+package gov.anl.aps.cdb.portal.model.db.entities;
+
+import gov.anl.aps.cdb.portal.constants.ItemDomainName;
+import javax.persistence.DiscriminatorValue;
+import javax.persistence.Entity;
+import javax.persistence.Table;
+
+/**
+ *
+ * @author djarosz
+ */
+@Entity
+@Table(name = "item")
+@DiscriminatorValue(value = ItemDomainName.LOCATION_ID + "")
+public class ItemDomainLocation extends Item {
+
+    @Override
+    public Item createInstance() {
+        return new ItemDomainLocation();
+    } 
+    
+}
